+++ src/NeuralNet/Layers/AlphaDropout.php
@@ -159,7 +159,7 @@ class AlphaDropout implements Hidden, Nonparametric
                 . ' backpropagating.');
         }
 
-        $errors = call_user_func($prevErrors)->hadamardProduct($this->mask);
+        $errors = $prevErrors()->hadamardProduct($this->mask);
 
         unset($this->mask);